Engine: skip storing raw sprites if TextureCache is close in capacity
This lets the engine to skip storing raw sprites in the SpriteCache when creating textures, if TextureCache is comparaible to SpriteCache in capacity, or larger. In such case there's little point in keeping the raw versions of sprites, as reloading a texture would involve reloading sprite itself anyway. If some sprites are already in cache (for whatever reason) they will not be removed of course.
If TextureCache is much smaller, then having a source sprite in the SpriteCache would save some time recreating a requested texture.

Note that raw sprites may still be loaded and kept when they are requested for software drawing, such as preparing GUI graphics or raw drawing (dynamic sprites, drawing surfaces).