Engine: fix use of global "walkspeed" variables in the pathfinder
These variables had to be set using set_route_move_speed() call in order for the find_route() and calculate_move_stage() work properly. But it's not obvious, and easy to forget.
So turn them into function arguments instead.

This fixes Character.AddWaypoint potentially using wrong walkspeed, if there are multiple characters or objects moving simultaneously.
An example of a script that would cause a bug:

    cEgo.AddWaypoint(20, 20);
    oObject0.Move(0, 160, 1, eNoBlock, eAnywhere);
    cEgo.AddWaypoint(200, 200);