Engine: removed global decl of "spriteset" from Common, fix gui init
Some operations in GUI require presence of spriteset, so I added a Spriteset pointer to GuiContext struct. This is not ideal, but will work for a short while.

Additionally, I had to remove "UpdateMetrics" call from ReadFromFile / ReadFromSavegame methods, as that requires access to runtime resources, such as sprites and fonts.
Instead UpdateMetrics is forced by sending OnResized event to all controls in a batch right after loading game data or restoring the save.
I think that there should be some kind of "Enter The Play" event for the game objects eventually that would do these things.