Engine: rewrote dynamic sprite to texture notification mechanism
The previous attempt, using a boolean spritemodified array, proved to be having logical loopholes. In simple words, because spritemodified flag was resetting each draw pass, it won't affect objects that are temporarily not visible, and they failed to reupdate when becoming visible again. This affected Software renderer.

Texture-based renderers have a different problem, where they fail to recognize when an old sprite was deleted and a completely new dynamic sprite created accidentally with the same ID. The cause was simply that the check for this was done inside sync_object_texture(), but sync_object_texture itself was not called unless draw() function thought that object had changed. So there has to be a check outside too...

Here I decided to reimplement notification mechanism, and use it for both types of renderers, for the sake of more straightforward code logic. (Although, strictly speaking, texture-based renderers could check their texture's "ref id" instead.)

The idea of a new implementation is this:
1. There's a map of "control blocks" per each active sprite (that is - sprite that was drawn as a part of object). The "control block" is simply a integer with a sprite ID in it, allocated and wrapped by shared_ptr.
2. When the ObjectTexture syncs its texture with a new sprite, the control block is retrieved (or new created), and from now on it may be shared among multiple objects which display same sprite.
3. When the sprite gets either updated or deleted two things happen: the previous control block is assigned an "invalid value", serving as a marker, and then it is removed from the map, left only if referenced by any ObjectTexture.
4. When checking whether texture needs to be re-synced we test the control block too, to see if it was invalidated. In such case this tells us to call sync_object_texture() again.

NOTE that dynamic sprites that do not appear on screen, and are only used for drawing on another surface, do not allocate control blocks, nor notify anything.