Editor: fix script editor fold in comment in scintilla lexxer
The issue here is if foldExplicitStart is empty it is defined as `"//{"`  and foldExplicitEnd it is defined as `"//}"`.

This causes the confusing behavior that if foldCommentExplicit  is true, the scintilla cpp lexxer can fold in comments following the previously defined start and terminator string. We are setting it to false to disable this unexpected behavior.