Common: calculate real TTF glyphs extent, fix fonts cut from the top
This is an attempt to fix certain fonts getting "cut off" from the top when being drawn on a limited size texture or bitmap.

The reason for such behavior is that fonts may include glyphs which topmost extent is higher than the reported "ascender".
In this case the engine simply does not know that glyphs may be drawn higher, and the prepared texture is not accommodated for that.

How we fix this:
1. Following a caller's instruction (a new flag ALFONT_FLG_PRECALC_MAX_CBOX), alfont preloads all glyphs one by one and records the maximal extents of their "control boxes".
These values are saved and returned on demand.
2. Our FontMetrics struct has a new VExtent field, that keeps these values (converted to top-down Y axis) for the reference.
3. get_font_surface_height() now returns a surface height calculated from these extent values.
4. CalcFontGraphicalVExtent() now uses the extent values when telling how the text will be graphically positioned, relative to the requested "text position".