Engine: reimplement OVR_AUTOPLACE, keep explicit overlay position
The purpose of this change is to have x,y parameters of ScreenOverlay to contain explicit coordinates at all times, and use coordinates strictly depending on the Overlay's layer (screen or room).

Historically, non-blocking character speech overlays stored a magic OVR_AUTOPLACE value in their x parameter, and character's index in y parameter. This told the engine to calculate this overlay's drawing position based on linked character's position.
Furthermore, for blocking speech the flag kOver_PositionAtRoomXY told to assume Overlay stores its position in room coordinates, even though it's an overlay on a screen layer.

This made the meaning of x,y parameters complicated, and contributed to discrepancy between values set to Overlay.X and Y properties, and values returned by reading them.

In this change:
* Magical OVR_AUTOPLACE value is removed.
* Overlay's flag kOver_PositionAtRoomXY is replaced with kOver_AutoPosition; - that is safe to do, because PositionAtRoomXY was ever only set along with OVR_AUTOPLACE.
* bgSpeechForChar parameter is renamed to speechForChar, and is assigned regardless of whether speech is blocking or non-blocking.
* Overlays now always store explicit position in x,y.
* Auto-positioning is performed during game update, and new x,y parameters are assigned.
* Overlay.X,Y properties now get and set same values, corresponding to overlay's layer (screen or room).