Engine: don't autopos blocking speech, and fixed reading old flags
Making blocking speech overlays autopositioned turned out to be a mistake, because they have a different position calculation depending on a script command used.
Replicating calculations for blocking speech in autoposition_overlay() would be difficult, because of the various branches in display_speech and display_main functions.
It would take a good refactor first, before all the conditions could be properly analyzed.

Additionally, my previous assumption of "room-relative pos" being equivalent to "autoposition" is incorrect, so I fixed flags upgrade from older saves.

NOTE: turns out kOver_PositionAtRoomXY was never saved for non-room overlays in practice. Because it was only ever applied to blocking speech, and game cannot be saved while blocking speech is on screen.
It was saved for room overlays though.