Common: when calculating TTF's extents, use first 128 glyphs
This fixes loading very big fonts, which may take literally seconds to calculate if we process each glyph.

This load mode is primarily purposed for backwards compatibility, as with new games authors may fix / replace the font if they have issues.
Because of that, it's not a good thing to pessimize loading times for games that don't require this autofix.