Common: when calculating TTF font's extent just use FT_Face->bbox
This apparently already stores precalculated maximal glyphs extents.
Testing this side-by-side with the previous solution shows that these bbox values are mostly matching the results of loading and measuring each glyph.
They are 100% matching in the "bad" fonts which I've been adding this calculation for. For some "good" fonts they sometimes have 1 pixel less offset, but are not less than the extent given by (0, height).

And obviously this is faster than loading and calculating glyphs.