Plugin API: implemented IAGSEngine.Log()
The original plan was for the engine to dynamically allocate an individual logging group for each plugin, which would allow to distinguish plugin messages by their group names.
But this is currently complicated by the Editor which does not support arbitrary group ids. So I'm leaving this in plans for the future.