Engine: hotfix certain platforms to not save "warnings.log" in cwd
There are platforms where this behavior is not suitable even for a Debug game build, because writing to current working dir will modify the game bundle, causing reaction from the operating system or app store.
Introduce a IsLocalDirRestricted() method to the platform driver class, use it when "warnings.log" is created to know whether we are allowed to write it into cwd, or should be using a safer user dir.

For the time being, do not impose this restriction on Windows and Linux.

NOTE: this is a temporary hotfix. For the proper solution we'd need to revise "warnings.log" feature and possibly use config settings instead.