Engine: added new custom properties to save games
AudioClips are not writing their custom properties to saves. They are considered a "game resource", and not part of the game's state.
This may be reviewed later, when we find good use cases for their custom properties, and it will become apparent that having them in saves is more convenient than reinitializing them in script after restoring a save.
Although hopefully this won't be necessary. Personally, I'd rather propose to reduce number of things written to game saves. E.g. don't save other resources, such as dynamic sprites.

If we must, eventually, then audio clip's runtime changes should not saved by "index", but use a script name for identification and matching, so that any change to a list of audio clips in game won't break older saves.