Fix for flood filling masks not working correctly with image padding.  Also minor optimization.
It looks like images can have a different _paddedWidth that needs to be taken into account when calculating where in the byte array we are.  The flood fill algorithm was using the normal width instead of the padded width, so it would basically always flood the entire screen since the pixels would be off.

Also, added a minor optimization.  In our inner loop were were checking image[i] == initial && image[i] != replacement.  The only time you could ever pass the initial check but fail the replacement check is if initial == replacement.  In that case you're just doing if (a == b && a != b), so we can just skip the entire function for cases where the initial and replacement values are the same.