Editor: remove "Copy walkable area mask to regions" room command
AFAIK this operation is related to importing very old game projects from early version of AGS which did not distinguish regions and walkable areas.
This has been obsolete and redundant for more than a decade, but operation is present on a toolbar, and makes it easy to trigger and spoil your room.

We might have a common copy/paste mask operation instead.