Editor: renamed GetBitmapForSprite functions to clarify what they do
There's a second variant of GetBitmapForSprite which is called with extra width and height parameters. Although called the same, it hides what it really does: converting original sprite to 32-bit color depth.
This is non-obvious, and may lead to incorrect use of a function, and unexpected results.

Renamed the first trivial variant to GetSpriteBitmap,
second variant to GetSpriteBitmapAs32bit.