AGS.Native: removed RGB <-> BGR sprite pixel swaps
While analyzing the code, I noticed that these conversions are mostly doubled, which means that pixels end up with their initial values.
After the swap is removed there's no visible difference in the Editor's previews for sprites, and compiled game looks as expected.

It's not clear when this swap became redundant, maybe this is remains of the v2.72 editor, or early v3.* editor where more of the engine code was shared with the editor directly.