Engine: potential optimization for String_IndexOf(), use precalc offset
This shares a use of precalculated character offset, previously implemented for String_GetChars().
The idea is that when counting a character offset for a utf-8 string, the program would record last found offset in string's meta header, and then start from that offset when searching for the greater character index next time (if possible).

Additionally:
* search for "end" offset from the found "start", since end will always be >= start;
* simply use header.Length if "end" is expected to be an end of the string.