Common: fixed DirectoryRecursiveIterator
The error was that when saving last subdirectory search position for the current level, it was saving a "find any entries" iterator instead, which is already past end at this point.
As a consequence, whenever returning from a subdir, it will look like all subdirs are already exhausted, and search is stopped.
Only first subdirectory and first subdirectories in it will be searched.