Imply `*` in import stmt iff` SCOPT_NOAUTOPTRIMPORT` is `false`
Introduce the compiler flag `SCOPT_NOAUTOPTRIMPORT` which is `false` by default.

Normally, whenever AGS code declares an object variable then imply that it is pointered.
However, in import declarations of non-autopointered varables only do so when the flag is `false`.

In legacy AGS, the engine autogenerates code such as `import Object object[20];` where `object` is meant to be an array of `Object` and not an array of `Object *`. In this legacy situation, the flag should be set so that the compiler generates the correct code for this.