Editor: remap default Character and GUI colors on new item creation
This is necessary because we have default values of color properties set in palette indexes. But also because it's not possible to set two default values for 8-bit and 32-bit modes.
So if we leave these defaults in 8-bit, we have to convert to 32-bit after an item is created and added in game.