Engine: in ags_simulate_keypress() only do TEXT event for "old mode"
This is because old mode uses TEXT events for running "on_key_press" for printable characters.
Passing TEXT event in the new mode will result in inconsistent behavior, as only limited number of characters (<128 indexes) will be reported correctly in "on_text_input", and there may be incorrect calls too for unicode chars.
So it's better to leave that for later. Maybe we should somehow use SDL2 internals for converting simulated key presses to textual representation instead of trying to handle this ourselves.