Engine: reimplement optional dirs as a vector, adjust "--runfromide"
This gets rid of 3 hardcoded options for optional data, audio and voice dirs, and lets provide any arbitrary number of additional data directories instead.
These dirs are paired with asset filter strings, so it's possible to define which assets should be looked for in them.

The "runfromide" syntax is now:

    --runfromide <INSTALLDIR> [OPT_PATH1:FILTER[,OPT_PATH2:FILTER[, ...]]]