Engine: File.ReadStringBack() supports both term and non-term strings
ReadStringBack() assumes the string is prepended by the int32 length value.
Because of that it has practically no difference for the engine whether there's a null terminator or not: we still read same amount of data.
So, allocate buffer as if read length value is a string's value without terminator (so buffer allocates +1 more),
and then let ScriptString::Create() to fixup terminator if one was not present within the size of the buffer.

After this ReadStringBack() can successfully read both strings that have and those that do not have null-terminator (assuming they have prepended int32 length value).

Additionally, this fixes a case when the provided buffer either does not contain null terminator, or contains unexpected ones in the middle.
This may happen both in case of a mistake in code, and in case a mistake in deserialized data.