Editor: search for event handlers, ignoring comments or string literals
Here's a problem: InsideStringOrComment relies on scintilla styling. But styling is not necessarily present at the time if we just loaded a document when zoom-in to a function.
So we force the styling in case it was not done here yet.
Style only the portion of a document from the last styled position to the position of the found match (this speeds things up a little).