Engine: fix RunDialog command spawning a nested dialog state
This fixes RunDialog script command (or Dialog.Start) spawning a nested dialog state if called within a dialog option script. Something that it should not be doing.
Instead, check if we are inside the dialog, and override a option script's return value, which dialog executor will handle upon receiving control back.