Engine: detect missing event handlers in script, remember and warn
Add Enabled field to the InteractionEvents's members. Record call result at runtime and disable event handlers that were not found in script.
Issue a warning for the first time when this happened.

NOTE: Room events will likely be reported again when returning to the same room, as the state of event handler is not saved when room gets unloaded.