Engine: allow to restore rooms with less script data
NOTE: it's not entirely clear at this point how to deal with rooms having less data, because we may only ever test 1 current room when restoring a save, but any other room states cannot be tested without loading room files, which is too much for a test's sake. So we have to delay the test until each particular room is loaded.

Something for a future consideration...