Compiler2: remade built-in DynamicArray.Length into a member function
Old way: the Length pseudo-attribute is substituted with a generated static function call. This works "out of the system", and requires to inject a special handling for this particular function call.

New way: instead of generating a bytecode for the static function call, we generate actual *attribute* "T[]::Length" for the given dynamic array type "T[]", then we either generate or find a previously generated function of a built-in name "__Builtin_DynamicArray::get_Length" that will substitute such attribute's getter.
Same function works as a getter for all "T[]::Length" attributes.
The further bytecode is generated normally, as if this was a real attribute declared in a normal way, with no special handling.

Had to modify "ParseStruct_Attribute2SymbolTable" method, created a "ParseStruct_MasterAttribute2SymbolTable" method that lets to optionally pass qualified getter or setter names, which it attaches to the created attribute.

Adjusted Compiler2 Tests correspondingly. As DynamicArray.Length this is now a object-function call, not a static function call.