Engine: fixed Wait function's negative time case for pre-3.6.0 API
Prior to 3.6.0 script API Wait function was treating negative values as "no time", but since 3.6.0 it treats it as "infinite time".
Also, old engines let overflow happen when assigning > INT16_MAX values to the wait counter (which is int16). In the old engine this resulted in immediate timer stop, by pure accident.