Direct3D: move window to primary display before toggling real fs
This is done in case the windowed mode was on another monitor at the time of the switch.
Direct3D renderer is currently programmed so that exclusive fs only works on primary monitor. Driver will init fullscreen properly itself, but engine may remain thinking that it's on secondary, and report incorrect info. Also switching back to windowed mode will use wrong monitor's resolution for the reference.