Fix bug (2): Attribute of function call result must be callable
In an expression, function calls can be first in a chain of '.' components or non-first. In either case, the part after the function call must not be skipped.

Also fix a small bug in a Googletest: A compile was supposed to fail due to trailing symbols after a constructor; but it failed because a function call didn't have the parameter specified.