Engine: don't put forks into LoadedInstances list
This list's only purpose is to serve as a lookup table when doing function call into other script (SCMD_CALLAS).
Because the CALLAS fixups register only primary instance's ID, the forks actually use the primary instances when doing function calls too (see "_runningInst" member).
For that reason the forks don't have much use to be in this table.

As a consequence, this DOUBLES the available slots for active scripts. In practice this means that there may be 256 user scripts as opposed to 128 previously.
(127 and 255 actually, if we count current room script always occupying a slot).