Engine: global DrawIndex helps resolving obj sorting with eq z-order
This is a quick solution for resolving objects with identical z-order, in which case we'd like to keep their relative sort persistent, and avoid random sorting "swaps".

Global DrawIndex is allocated for each drawable object in game (except for gui controls that are sorted only among each other).
Static objects are reserved draw indexes at the game init.
Dynamic objects (currently only Overlays) have one assigned for them when they are *drawn for the first time*.

DrawIndex replaces use of GUI.ID and Overlay's "creation_id" for this purpose.