Fix bug: Handle empty 'std::string' properly
A `char[]` is considered an empty string whenever a '\0' is in its first byte, but a `std::string` cannot be handled that way: → use `clear()` and `empty()`.

Some strings in the interface of the compiler have been converted from `char[]` into `std::string`. Update the handling within the parser to account for that.