Revamp bytecode library functions
Recently, a lot of `char[]` in the interface to the compiler have been converted to `std::string`. Rewrite the functions that generate bytecode tests or that compare interface components to match. Also provide code to compare and check the `scrip.globaldata` that the compiler generates.

Make some library functions `static`, as suggested by the MSVS compiler.