Eliminate duplicate tests when function parameters are defined
When a parameter is defined in a function header and a body follows then that parameter can hide or collide with global definitions. Code for these checks is already provided where the compiler processes local variable defns. No need to duplicate the checks where parameter definitions are processed.

Fixes a bug : These checks were even done when a function body doesn't follow (import or forward declaration of a function), which is incorrect.

Provide googletests that demonstrate that the bug is fixed.