Compiler (old): support module names in script markers, use in RTTI
Compiler parses an optional appendix to the new script marker, which defines a "module name". This module name is used when compiling RTTI table, all types declared within a section are assigned to that "module", instead of a section.
When no module name is defined, compiler falls back to using section name instead (which is usually a name of a script file).