Engine: implement kScValPluginArgPtr script value type, for 64-bit ptrs
kScValPluginArgPtr is a type for RuntimeScriptValue that is used when we are sure that the value returned from plugin is a pointer, and not any number.
At the moment engine does not have normal means to deduce that, since script function prototypes are not being registered in the compiled script data.
But there's something that we still may do in 64-bit enviroment: knowing that AGS script only supports 32-bit numeric values, would we receive any value that exceeds 32-bit number, we may assume that's a pointer.

RuntimeScriptValue::SetPluginArgOrPtr() factory method checks whether provided intptr_t value exceeds 32-bits. If it does, then it assigns a kScValPluginArgPtr type, if it does not then it assigns kScValPluginArg type (can be a small pointer or a plain number).

Use SetPluginArgOrPtr() when receiving arguments in IAGSEngine::CallGameScriptFunction().