Engine: use SetPluginArgOrPtr() when getting plugin's return value
This gets rid of ccInstance::_pluginReturnValue hack, and instead uses a simpler solution, which is also a hack but perhaps less so.
On a good side the new workaround fixes the potential situation when the plugin creates several dynamic objects in a row, but returns not the last created one. The "plugin return value" solution would be unreliable in such case.