Engine: added DisplayTextPosition to DisplayTextLooks
This amends kDisplayTextStyle_Overchar with a more elaborate combination of settings.
Stop using negative x and y in display_main as a telling that the text should be centered.
(They are still used in display_speech though, since the script api supports this behavior, whether intentionally or by accident.)

Also fixes SayAt being clamped to the screen as if it's regular Say.