Engine: implement "direct" flag per MoveList's stage
This fixes the ancient bug related to AddWaypoint(), which causes character to teleport to a nearest walkable area if AddWaypoint follows Walk(eWalkableAreas), because then the MoveList is tagged as "non direct", and engine snaps character to a walkable area when stopping it.
The new implementation has flags per each stage, where stages created by Walk would have their own (either direct or non-direct), and any new stage added with AddWaypoint wll have a "direct" flag.
The engine will refer to the current or last stage's flag when deciding what to do.