Engine: explicitly restrict ground interactions during blocking actions
The problem in AGS engine is that it schedules interaction scripts to run after the current script completes. Which, in case of a blocking action (which still runs game updates on background), may occur any time later.
As a result, a lot of redundant events may accumulate, and run all at once, when they no longer make sense.

Previously, this was accidentally prevented by limiting scheduled callbacks to an arbitrary number of 4, where new scheduled item exceeding this count would overwrite the previous one.
This was solving the issue most of the time, but was still quite unreliable.

This change explicitly prevents any ground interaction events to trigger while there's a blocking action in a process.