Editor: use original GIF palette when using pal index as transparency
This is required, because when GifDecoder converts each frame to a Bitmap, the attached palettes may not be equal to the original image file's palette. They may be less in size, or have shuffled colors.
In order to work around that we query original palette and use it as a reference when:
- displaying selected color preview in SpriteImportWindow
- adjusting transparency index before calling internal sprite conversion method.