Engine: fix solid char gets ignored when inside walking char's zone
For some reason engine was testing 2 conditions:
- walking character is inside other char's blocking rect;
- other char is inside walking character's blocking rect.

If any of the above is true, then the "other char" was disabled from the obstacles list.

In effect, if the walking character's blocking rectangle is large enough compared to the standing character's blocking rectangle, then the character will be able to pass through.

I believe that's illogical, so removed the second test.