Engine: cache shader constants data in script objects
This caches the constants values in script shader and shader instance objects, independently from the real shaders.
This is done in case the real shader is not available in this run session or has mistakes, which will be fixed afterwards.
The cached data may be used when saving game and restoring later with a different graphics driver or a different configuration of shader assets.