Engine: script Shader objects serialization in saves
1. Added shader fields to all game object save formats.
2. ScriptShaderProgram and ScriptShaderInstance serialize along with the managed pool.
3. Prior to restoring a save all shader instances are deleted by a gfx driver (because ScriptShaderInstance get disposed). Shaders are kept, as they may be reused by a restored save.
4. Upon restoring a save the game recreates all shaders and shader instances in gfx driver. Shaders that still exist do not get recreated again, of course. Shader instance get their constant data reapplied.

The idea is that it should not matter which gfx driver was used when the save was made, and which is used when it's restored. The shader data is saved and reapplied regardless by the script shader objects.
Actual shaders may or not be working of course, depending on current runtime setup.