Engine: consistent translation: only translate a text when displaying
This removes inconsistent translation of game texts when they are either assigned to object properties or retrieved from ones.
The principle to follow is this: properties set and get untranslated strings, the text is only automatically translated when displayed or drawn somewhere.
Note: user still has an option to explicitly translate anything using GetTranslation function.