Engine: special "post fx" shader setup for the final surface render
This adds a "Output Dimensions" constant, but it's set only for the final surface render, so suitable only for the Screen shader.
The values are assigned depending on the current backbuffer state, and will vary between an output to the window and an output to the texture (in case of e.g. making a screenshot).