Editor: introduced timeout for "Check for updates" process
NOTE: at first I tried using WebClient.DownloadDataAsync and do downloading asynchronously, but then it turned out that this call hangs the calling thread in case there's any problem with the connection or a proxy.
Which means that this call itself has to be on timeout... or another API used for establishing connection overall.