Common: increment game data version as a safety fallback
There are few fixes in text alignment math done in the recent patches.
It's hard to predict whether they will cause any regression in looks (noticeable regressions) in the existing games, but just in case they will, we could use this new data version tag to detect a change in behavior.