Engine: if game's font file is missing, init the font with the file 0
This was an old behavior to replace a missing WFN font with WFN 0 implicitly (previously it was done inside the WFNRenderer).
Had to reimplement it in a new way, because some old games seem to be missing fonts.
Also this will at least display a text in case some file was not packed.