Engine: when starting to move wait 1 frame to let animation advance
This reimplements a default AGS behavior where character will keep at the previous position for 1 frame more before starting to move along the path.
Previously this worked because MoveList returned a *previous* position when updating. It is now returning a *new* position after a MoveList refactor and a introduction of MotionPath api. Therefore the "animation first" behavior has to be implemented in another way.
IMHO it's reasonable to not have this logic inside MoveList, as that makes MoveList having inconsistent state; instead this logic shall be a part of the Character which decides when to animate and when to move along the path.
Users of the MotionPath api can do according to their desire in script.