Editor: implemented UpgradeGameManager
UpgradeGameManager is a class that acts as a controller for all the tasks performed when upgrading an older game project.
It gathers a collection of Upgrade Tasks, some of which may require to display a cue to user about a bigger change to the project, others are meant to be executed implicitly and without condition.
If there are any tasks that imply a big change, then UpgradeGameManager creates and shows a "Upgrade Wizard" dialog.

IUpgradeGameTask is a interface of a single upgrade task (although a single task may perform various updates, how they are organized is a matter of convention). This interface provides means to know which game versions should have this task applied, and instructions on how to handle this task (whether it may be canceled by user, skipped in case of a failure, and so forth).
IUpgradeGameTask also has a method that creates a dialog page for the "Upgrade Wizard", with information and related options (if necessary).

UpgradeGameManager has its own list of "common" upgrade tasks which it creates itself, but also fires a EditorEvents.GamePrepareUpgrade event, which lets any Editor's component to add its own Upgrade Task.