Editor: hide certain properties when multiple objects are selected
Implemented BrowsableMultieditAttribute that marks properties that should or should not be displayed whenever multiple objects are selected for edit in PropertyGrid.
When PropertiesPanel.SelectedObjects is set, check whether there are multiple objects, and if yes then assign this custom attribute to the propertiesGrid.BrowsableAttributes collection.

This allows to hide properties that must have unique values, such as Name, so that user won't try to edit them.