Editor: add Go To command in Project Explorer nodes
This adds a Go To Dialog that can be opened from the top level of the Audio Clip, Character, Dialog, GUI, Inventory Item, Room and View nodes.

This dialog is generalized so it can be reused across these elements. It's also meant to be quick and easy to use with either keyboard or mouse input.