Common: backwards compatible behavior of GUIListBox's selection
Prior to 3.6.3 GUIListBox secretly kept selection on index 0 even if there were no items yet, or if list box was Cleared (but NOT when items were Removed one by one).
This resulted in selection appearing on the first item automatically whenever one was added at start or after pre-clearing a list. (Usually this mattered in "save/restore" menus.)