Engine: fixed 16-bit games had default bitmap depth reset to 32-bit
This mistake apparently dates back the v3.6.0 when we switched to SDL2 backend, and rewrote Software renderer to draw native game screen to a 32-bit texture before displaying.
Somehow we missed that the graphics driver was still changing default bitmap color depth to the final display depth, which it should not be doing.
This caused various strange effects in 16-bit games when player toggled the display mode mid-play.