Editor: consistent Script Name and Description/Name property labels
This assigns consistent labels to "script name" and "description" kind of properties in all the game entities. This also makes them consistent with how they are called in script api.
Only DisplayName of properties are changed, leaving property names as they are, so there should not be any compatibility issues and data format is not changed.
The effect is purely in how the users will see these properties are called in the editor panel.